LPS-69043 Move non-exported packages to "internal"
