+++ modules/util/bundle-support/src/main/java/com/liferay/bundle/support/BundleSupport.java
@@ -16,7 +16,7 @@ package com.liferay.bundle.support;
 
 import com.beust.jcommander.JCommander;
 
-import com.liferay.bundle.support.util.FileUtil;
+import com.liferay.bundle.support.internal.util.FileUtil;
 
 import java.io.File;
