+++ modules/util/bundle-support/src/main/java/com/liferay/bundle/support/CommandClean.java
@@ -17,7 +17,7 @@ package com.liferay.bundle.support;
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
 
-import com.liferay.bundle.support.util.FileUtil;
+import com.liferay.bundle.support.internal.util.FileUtil;
 
 import java.io.File;
 import java.io.IOException;
