+++ modules/util/bundle-support/src/test/java/com/liferay/bundle/support/BundleSupportTest.java
@@ -14,7 +14,7 @@
 
 package com.liferay.bundle.support;
 
-import com.liferay.bundle.support.util.FileUtil;
+import com.liferay.bundle.support.internal.util.FileUtil;
 
 import com.sun.net.httpserver.BasicAuthenticator;
 import com.sun.net.httpserver.Headers;
